// src/api/resources/voices/types/ModelEntity.ts
export interface ModelEntity {
    _id: string;
    type: "svc" | "tts";
    title: string;
    description: string;
    cover_image: string;
    state: "created" | "training" | "trained" | "failed";
    tags: string[];
    created_at: string;
    updated_at: string;
    visibility: "public" | "unlist" | "private";
    like_count: number;
    mark_count: number;
    shared_count: number;
    task_count: number;
    author: {
        _id: string;
        nickname: string;
        avatar: string;
    };
    train_mode?: "fast" | "full";
    samples?: {
        title: string;
        text: string;
        task_id: string;
        audio: string;
    }
    languages?: string[];
    lock_visibility?: boolean;
    unliked?: boolean;
    liked?: boolean;
    marked?: boolean;
}